+++ src/app/app.module.ts
@@ -76,7 +76,7 @@ import { SettingsZonesComponent, DialogNewZone } from './settings-zones/settings
 import { WidgetIframeComponent } from './widget-iframe/widget-iframe.component';
 import { SettingsConfigComponent } from './settings-config/settings-config.component';
 
-import { GaugesModule } from '@biacsics/ng-canvas-gauges';
+import { GaugesModule } from './gauges-module/gauges.module';
 import { WidgetGaugeNgLinearComponent } from './widget-gauge-ng-linear/widget-gauge-ng-linear.component';
 import { WidgetGaugeNgRadialComponent } from './widget-gauge-ng-radial/widget-gauge-ng-radial.component';
 import { AlarmMenuComponent } from './alarm-menu/alarm-menu.component';